requirements.txt
openai
sentence-transformers
pdfplumber
python-docx
python-dotenv
rapidfuzz